64 bit compatibility
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@2348 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/config.c
@@ -687,7 +687,8 @@ void module_index(void)
 	/* Set up a unique job identity as a 4-bytes integer */
 unsigned long create_job_ident(void)
 {
-    unsigned long it;
+/* unsigned long it; // 64-bit compatibility: take away _long_ */
+    unsigned it;
     int i;
     	/* Is this enough to guarantee uniqueness? */
